Consolidate graph variants and declare runner defaults
